requirements.txt
langchain-groq
python-decouple
